requirements.txt
pillow==11.2.1
numpy==2.2.5
matplotlib==3.10.1